Remove default impl for `Iterator<Item=Input>`.
This turned out to make debugging rule implementations much harder,
because Rust would complain about "T is not an iterator" rather than
"no CodeMakerRule impl for T". Instead, we have a `make_from_iter`
helper method on the trait.

This has the added advantage of removing the lifetime generic in
the `CodeMakerRule` trait, dramatically simplifying the macro that
we use to define it.