Experiment with quoting python syntax inline in rule definitions.
The idea here is to let you write templated python(ish) syntax inline in
the definition of a codemaker rule that targets python, by using a proc
macro that translates Rust syntax tokens into a series of codemaker_python
builder API calls.

I think it has potential, but there are a lot of fiddly details. This will
probably only be tractable if we can abstract the core of the idea into the
main "codemaker" crate to simplify implementation for each backend.

I think it's promising, but there